requirements.txt
pact-python==0.18.0
pytest==4.4.1
Flask==1.0.2
